feat: content browser icons

+++ SolutionShelves/src/Platform/OpenGL/OpenGLTexture.h
@@ -15,7 +15,7 @@ namespace SolutionShelves
 
 		virtual uint32_t GetWidth() const override { return m_Width; }
 		virtual uint32_t GetHeight() const override { return m_Height; }
-		virtual uint32_t GetRendererID() const override { return m_RendererID; }
+		virtual uint64_t GetRendererID() const override { return m_RendererID; }
 
 		virtual void SetData(void* data, uint32_t size) override;
 
@@ -31,7 +31,7 @@ namespace SolutionShelves
 		std::string m_Path;
 		bool m_IsLoaded = false;
 		uint32_t m_Width, m_Height;
-		uint32_t m_RendererID;
+		uint64_t m_RendererID;
 		GLenum m_InternalFormat, m_DataFormat;
 	};
 }
